requirements.txt
# MCP Server
mcp[cli]>=1.0.0
starlette>=0.35.0
uvicorn[standard]>=0.27.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Git Operations
GitPython>=3.1.40

# Semantic Search
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# OAuth / Auth
PyJWT>=2.8.0
cryptography>=41.0.0

# Logging
structlog>=24.1.0

# HTTP Client
httpx>=0.25.0
